Attempt to modify cypress firmware for new Python 480 sensor
The imaging area of the ON Semi Python 480 differs from that of the Teledyne e2v
EV76C541, and the USB descriptors of the DAQ board must be changed to reflect
this difference. The resolution of the Teledyne EV76541 is 752 X 480, while the
Python 480 is 808 X 608. Both the format descriptor and frame descriptor need to
be modified, in both the regular and super speed USB descriptors.

The format descriptor includes a field for the aspect ratio, which for 808 x 608
is 101:76. In hex, this ratio is 0x65:0x4C.

The frame descriptor includes a field for the pixel sizes of the video feed,
which is 808 by 608 or in hex 0x0328 by 0x0260. Because the USB descriptors are
little endian, this is encoded as 0x28, 0x03 and 0x60, 0x02.