Disable hue control of processing unit
The Processing Unit descriptor, described in the UVC format, previously
indicated that 'hue' control was available for this UVC device. This channel was
re-purposed to control the LED intensity of the last version of the
feather-scope. Since the illumination is no longer being provided by an on-board
LED, and is instead now being provided by a laser outside of the endoscope, this
control is no longer required.

I have removed the hue control from the USB descriptor to reflect this change in
the microscope design, and have additionally removed the code to handle hue
control requests on the control endpoint.